Upgrade @rails/webpacker
When trying to build in Docker, it fails due to looking for Python 2.x; it appears this is a dependency of a package called node-gyp, which we're using a version of from 2018 (though Python 2.x was very old even in 2018). node-gyp in turn is a dependency of node-sass, which is a dependency of @rails/webpacker. So this updates that — it's only actually a minor version change, so shouldn't(!) break too much.