Use the new headless chrome
The selenium-webdriver gem mis-identifies the old headless chrome as
is it now identified as 'chrome-headless-shell' and the correct
extensions are not loaded. This should be fixed in a new release of
the gem but for now opt in to using the new version. There are reports
of the new version being slower and more resource intensive so review
once the new gem is out and compare performance between the two modes.